feat(suite-native): show info about sell error, add test case
